Implémentation des etats de saut random + des probas sur les actions des transitions + Nouveaux automates pour faire des déplacements random

+++ game/controller/condition/tree/BotFunCall.java
@@ -8,10 +8,10 @@ public class BotFunCall implements ICondition {
 	public String m_name;
 	public String m_p1;
 	public String m_p2;
-	public int m_percent;
+	public float m_percent;
 
 
-	public BotFunCall (String name, String p1, String p2, int percent) {
+	public BotFunCall (String name, String p1, String p2, float percent) {
 		m_name = name;
 		m_p1 = p1;
 		m_p2 = p2;
